+++ src/sys/mem/borrow/mod.rs
@@ -3,7 +3,7 @@
 //! Borrowed data.
 //
 
-mod alloc; // AllocMode
+mod backing; // Backing
 mod maybe; // MaybeOwned
 mod ownership; // Ownership, BackingChoice
 mod reexports;
@@ -13,7 +13,7 @@ crate::items! { // structural access: _mods, _all
     pub use _mods::*;
 
     mod _mods { #![allow(unused)]
-        pub use super::{alloc::*, maybe::*, ownership::*, reexports::*};
+        pub use super::{backing::*, maybe::*, ownership::*, reexports::*};
     }
     pub(super) mod _all { #![allow(unused)]
         #[doc(inline)]